expense-report-system/backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
asyncpg>=0.28.0
python-dotenv>=1.0.0
pydantic>=1.10.0
redis>=4.6.0
